refactor: styles for AssetIcon
