+++ packages/components/asset/package.json
@@ -9,6 +9,8 @@
     "@babel/runtime": "^7.6.2",
     "@contentful/f36-core": "^4.0.0",
     "@contentful/f36-tokens": "^4.0.0",
+    "@contentful/f36-icons": "^4.0.0",
+    "@contentful/f36-icon": "^4.0.0",
     "emotion": "^10.0.17"
   },
   "peerDependencies": {
