--- packages/components/asset/src/AssetIcon/AssetIcon.css
@@ -1,11 +1,0 @@
-.AssetIcon {
-  display: inline-block;
-  /* todo: should we just create a custom size property for Icon component */
-  width: 39px !important;
-  height: 39px !important;
-}
-
-.AssetIcon g,
-.AssetIcon path {
-  fill: var(--color-text-light);
-}
