+++ packages/components/asset/src/AssetIcon/AssetIcon.tsx
@@ -13,9 +13,10 @@ import {
   VideoIcon,
 } from '@contentful/f36-icons';
 import type { GeneratedIconProps } from '@contentful/f36-icon';
-import { AssetType } from '../Asset';
-import styles from './AssetIcon.css';
-import cn from 'classnames';
+import type { AssetType } from '../types';
+
+import { cx } from 'emotion';
+import { getAssetIconStyles } from './AssetIcon.styles';
 
 export interface AssetIconProps
   extends Omit<GeneratedIconProps, 'illustration' | 'ref'> {
@@ -30,9 +31,10 @@ export function AssetIcon({
   className,
   ...otherProps
 }: AssetIconProps) {
+  const styles = getAssetIconStyles();
   const props = {
     ...otherProps,
-    className: cn(className, styles['AssetIcon']),
+    className: cx(styles.root, className),
   };
 
   switch (type) {
