+++ packages/components/asset/src/index.ts
@@ -1,4 +1,6 @@
-export { Asset, isAssetType } from './Asset';
-export type { AssetProps, AssetStatus, AssetType } from './Asset';
+export { Asset } from './Asset';
+export type { AssetProps, AssetStatus } from './Asset';
 export { AssetIcon } from './AssetIcon/AssetIcon';
 export type { AssetIconProps } from './AssetIcon/AssetIcon';
+export type { AssetType } from './types';
+export { isAssetType } from './types';
